fix argc,argv issue

+++ Sources/Basic/Qt/qclassifier.cpp
@@ -21,7 +21,7 @@ void QClassifier::classifyImage(qintptr _taskid, const QByteArray &_encimg)
 {
     QJsonObject _json;
     cv::Mat _img = cv::imdecode(std::vector<unsigned char>(_encimg.begin(),_encimg.end()),cv::IMREAD_UNCHANGED);
-    std::vector<double> _vconf;
+    std::vector<float> _vconf;
     int _error = 0;
     ptrrec->predict(_img,_vconf,&_error);
     if(_error == 0) {
